requirements.txt
asgiref==3.11.0
Django==5.2.9
django-allauth==65.13.1
sqlparse==0.5.5
typing_extensions==4.15.0
